Properly fixed issues around blocking access to things when automation is running

+++ app/Game/Factions/FactionLoyalty/Controllers/Api/FactionLoyaltyController.php
@@ -5,6 +5,8 @@ namespace App\Game\Factions\FactionLoyalty\Controllers\Api;
 use App\Flare\Models\Character;
 use App\Flare\Models\Faction;
 use App\Flare\Models\FactionLoyaltyNpc;
+use App\Game\Automation\Concerns\ChecksAutomationRestrictions;
+use App\Game\Automation\Services\AutomationRestrictionService;
 use App\Game\Factions\FactionLoyalty\Services\FactionLoyaltyService;
 use App\Http\Controllers\Controller;
 use Exception;
@@ -12,6 +14,8 @@ use Illuminate\Http\JsonResponse;
 
 class FactionLoyaltyController extends Controller
 {
+    use ChecksAutomationRestrictions;
+
     private FactionLoyaltyService $factionLoyaltyService;
 
     public function __construct(FactionLoyaltyService $factionLoyaltyService)
@@ -35,6 +39,11 @@ class FactionLoyaltyController extends Controller
      */
     public function pledgeLoyalty(Character $character, Faction $faction): JsonResponse
     {
+        $restriction = $this->automationRestrictionJsonResponse($character, AutomationRestrictionService::START_FACTION_LOYALTY);
+
+        if (! is_null($restriction)) {
+            return $restriction;
+        }
 
         $response = $this->factionLoyaltyService->pledgeLoyalty($character, $faction);
 
@@ -46,6 +55,11 @@ class FactionLoyaltyController extends Controller
 
     public function removePledge(Character $character, Faction $faction): JsonResponse
     {
+        $restriction = $this->automationRestrictionJsonResponse($character, AutomationRestrictionService::START_FACTION_LOYALTY);
+
+        if (! is_null($restriction)) {
+            return $restriction;
+        }
 
         $response = $this->factionLoyaltyService->removePledge($character, $faction);
 
@@ -57,6 +71,12 @@ class FactionLoyaltyController extends Controller
 
     public function assistNpc(Character $character, FactionLoyaltyNpc $factionLoyaltyNpc): JsonResponse
     {
+        $restriction = $this->automationRestrictionJsonResponse($character, AutomationRestrictionService::START_FACTION_LOYALTY);
+
+        if (! is_null($restriction)) {
+            return $restriction;
+        }
+
         $response = $this->factionLoyaltyService->assistNpc($character, $factionLoyaltyNpc);
 
         $status = $response['status'];
@@ -67,6 +87,12 @@ class FactionLoyaltyController extends Controller
 
     public function stopAssistingNpc(Character $character, FactionLoyaltyNpc $factionLoyaltyNpc): JsonResponse
     {
+        $restriction = $this->automationRestrictionJsonResponse($character, AutomationRestrictionService::START_FACTION_LOYALTY);
+
+        if (! is_null($restriction)) {
+            return $restriction;
+        }
+
         $response = $this->factionLoyaltyService->stopAssistingNpc($character, $factionLoyaltyNpc);
 
         $status = $response['status'];
